Refactored index to use api_view decorator
