+++ test/views/unit/index_get_tests.py
@@ -1,11 +1,10 @@
-import json
 from rest_framework import status
 
-from backend.controllers.index import index
+from backend.views.index import index
 
 
-def test_index_returns_status():
-    request = {}
+def test_index_returns_status(request_builder):
+    request = request_builder("GET", "/")
     response = index(request)
-    assert json.loads(response.content) == {"status": "running"}
     assert response.status_code == status.HTTP_200_OK
+    assert response.data == {"status": "running"}
